refactor: [FE] 에디터 컴포넌트 분리 에디터 컴포넌트 분리
TODO: 반응형
(#71)
