+++ frontEnd/src/components/room/EditorSection.tsx
@@ -1,4 +1,4 @@
-import Editor from '../Editor';
+import Editor from './editor/Editor';
 
 export default function EditorSection({
   dataChannels,
